Updated Tests and fixed continue
Updated the tests on Bb. The only one giving me a really hard time is
5... it just doesnt work correctly. Tests 8 and 9 should give back
errors, but the layers code needs to be looked at a bit further because
its not being popped off properly. Also, the cps for while needs to be
written... mine isnt which is why i didnt include test 5 in the test
file.